requirements.txt
Flask
flask-cors
gunicorn
python-dotenv
mysql-connector-python
pymysql
requests
